+++ examples/data-sources/meraki_appliance_organization_security_intrusion/data-source.tf
@@ -1,0 +1,3 @@
+data "meraki_appliance_organization_security_intrusion" "example" {
+  organization_id = "123456"
+}