+++ examples/resources/meraki_appliance_organization_security_intrusion/import.sh
@@ -1,0 +1,1 @@
+terraform import meraki_appliance_organization_security_intrusion.example "<organization_id>"